Move utility functions to utils.py
Refactored sanitize_list, flatten_dict, and format_appliance_id_for_api from models.py to utils.py for better code organization and reuse. Updated imports and removed duplicate code from models.py.
